Finish basic servo controls
Wraps up preliminary implementation of user provided servo commands. The
intent is that the keyup and keydown presses will send a json string to
the server, which can be parsed into a simple struct and used to command
the motors.

Uses two functions, servo_control and send_command, which determine the
servo state and send the correct value to the gpio daemon respectively.